cleanup build warnings 1. Change `CyaSSL_OCSP_set_options()` to return `SSL_SUCCESS`    or `SSL_FAILURE` as `int` like rest of API. 2. Fix data narrowing warning in file io.c function    `process_http_response()`. 3. Fix global variable shadowed warning in file ssl.c function    `CyaSSL_GetSessionAtIndex()` 4. Fix data narrowing warning in file internal.c functions    `Encrypt()` and `Decrypt()`. Passed in a word32 size parameter    that was provided a word16 and used as a word16. 5. Removed unreachable code from file tls.c function    `CyaSSL_GetHmacType()`. 6. Fix data narrowing warnings in file aes.c functions    `AesCcmEncrypt()` and `AesCcmDecrypt()`.

+++ cyassl/ssl.h
@@ -1140,7 +1140,7 @@ CYASSL_API int CyaSSL_accept_ex(CYASSL*, HandShakeCallBack, TimeoutCallBack,
 #endif /* CYASSL_CALLBACKS */
 
 
-CYASSL_API long CyaSSL_CTX_OCSP_set_options(CYASSL_CTX*, long);
+CYASSL_API int  CyaSSL_CTX_OCSP_set_options(CYASSL_CTX*, int);
 CYASSL_API int  CyaSSL_CTX_OCSP_set_override_url(CYASSL_CTX*, const char*);
 
 /* OCSP Options */
